ROS parameters cannot be floats.
git-svn-id: https://atrias.googlecode.com/svn/trunk@2122 a504bc44-2baa-4cb8-4136-ba805e94d278

+++ software/atrias_controllers/atc_leg_position/include/atc_leg_position/controller_gui.h
@@ -35,13 +35,16 @@ Gtk::HScale *leg_length_hscale,
 Gtk::CheckButton *update_checkbutton;
 
 // Parameters
-float leg_length_param;
-float leg_angle_param;
-float leg_p_gain_param;
-float leg_d_gain_param;
-float hip_angle_param;
-float hip_p_gain_param;
-float hip_d_gain_param;
+//     NOTE: ROS parameters cannot be floats so loss-of-precision issues do not
+//     arise when both C and python nodes access the same parameter. See:
+//     http://answers.ros.org/question/10938/why-cant-you-use-floats-for-accessing-parameters-in-roscpp/
+double leg_length_param;
+double leg_angle_param;
+double leg_p_gain_param;
+double leg_d_gain_param;
+double hip_angle_param;
+double hip_p_gain_param;
+double hip_d_gain_param;
 
 void controllerCallback(const atc_leg_position::controller_status &status);
